custom-nodes/node-red-contrib-emotiondetectionnode/requirements.txt
opencv-python
torch
torchvision
Pillow
numpy
onnxruntime
future